requirements.txt
accelerate==1.13.0
aiofile==3.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==26.1.0
Authlib==1.6.9
backports.tarfile==1.2.0
beartype==0.22.9
brotli==1.2.0
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chess==1.11.2
click==8.3.1
comm==0.2.3
cryptography==46.0.5
cuda-bindings==13.2.0
cuda-pathfinder==1.4.4
cuda-toolkit==13.0.2
cyclopts==4.10.1
datasets==4.8.4
debugpy==1.8.20
decorator==5.2.1
dill==0.4.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.2
fastmcp==3.1.1
ffmpy==1.0.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==8.38.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.4
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.29.0
openapi-pydantic==0.5.1
openenv-core==0.2.3.dev0
-e git+https://github.com/meta-pytorch/OpenEnv.git@ca8e7c0970ebd2c2623248e58dedf9664fb5ae21#egg=openenv_echo_env&subdirectory=envs/echo_env
-e git+https://github.com/meta-pytorch/OpenEnv.git@ca8e7c0970ebd2c2623248e58dedf9664fb5ae21#egg=openenv_openspiel_env&subdirectory=envs/openspiel_env
-e git+https://github.com/meta-pytorch/OpenEnv.git@ca8e7c0970ebd2c2623248e58dedf9664fb5ae21#egg=openenv_textarena&subdirectory=envs/textarena_env
opentelemetry-api==1.40.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
parso==0.8.6
pathable==0.5.0
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.4.4
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
SecretStorage==3.5.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.3
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
TextArena==0.7.4
tokenizers==0.22.2
tomli==2.4.0
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
trl==0.29.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
